increase influence of gravity/fix calc. for Pluto
When the numbers get very small, which is the case when calculating the
numbers for Pluto, odd things can happen. For instance, the light might
end up 180 degrees set off from where it should be. Now, this doesn't
happen with every build I make. Since the only differences from one
build to another are the accelerator and gyroscope offsets, I assume,
that this might play a role in it. As far as I can tell, when the
gyroscope offsets are above the usual offsets, the odd things start
happening. To mitigate this we increasing the influence of gravity. I
strongly assume that cheap MPU6050 breakout boards have a too high
offset and therefore cause this issue.

Increasing GRAV_MULTI slightly makes the pendulum heavier but seems to
fix the issue, as far as I can tell.