loosen on/off mechanism
New 3d printers with input shaping exhibit slightly different print
dimensions. This can lead to very tight tolerances around the on/off
mechanism. The design is already made so that the mechanism still works
even if the mechanism is loose. Therefore loosen the mechanism so much
so that if printed on a prusa mk4 with input shaping the mechanism
doesn't lock in place. After printing the loosened version on mk4, I no
longer had to post process the parts, which is a nice side effect of
this change.